+++ src/harlequin_duckdb/cli_options.py
@@ -26,12 +26,6 @@ no_init = FlagOption(
     description="Start Harlequin without executing the initialization script.",
 )
 
-read_only = FlagOption(
-    name="read-only",
-    short_decls=["-readonly", "-r"],
-    description="Open the database file in read-only mode.",
-)
-
 unsigned = FlagOption(
     name="allow-unsigned-extensions",
     description="Allow loading unsigned extensions",
@@ -79,7 +73,6 @@ md_saas = FlagOption(
 DUCKDB_OPTIONS = [
     init,
     no_init,
-    read_only,
     unsigned,
     extensions,
     force_extensions,